requirements.txt
# Pinned for Python 3.12 target; also build on 3.13/3.14 via floor ranges.
fastapi>=0.115,<0.118
uvicorn[standard]>=0.34,<0.36
sqlalchemy[asyncio]>=2.0.40,<2.1
asyncpg>=0.30,<1.0
redis[hiredis]>=5.2,<6.0
python-jose[cryptography]>=3.4,<4.0
jsonschema>=4.23,<5.0
httpx>=0.28,<0.29
pydantic>=2.11,<3.0
pydantic-settings>=2.6,<3.0
daytona-sdk>=0.102,<1.0

# ─── Test ───
pytest>=8.3,<9.0
pytest-asyncio>=0.24,<1.0
pytest-httpx>=0.35,<1.0
aiosqlite>=0.20,<1.0
